Потребовалась замена unsigned на int в некоторых местах, с целью возможности проверки передачи в качестве параметров отрицетальных значений

+++ neural_network/neural_network_lib/include/layer.h
@@ -7,7 +7,7 @@ class Layer
 {
 public:
     Layer() = delete;
-    Layer(const size_t& kInputsNumber, const size_t& kNeuronsNumber,
+    Layer(const int& kInputsNumber, const int& kNeuronsNumber,
           const ActivationFunctionType& kActivationFunction);
 
     void InitializeRandomWeights(const double& min_value, const double& max_value);
